// src/app/pages/04-bodega/interfaces/print-kardex..ts
export interface BodegaPrintKardexInterface {
    Ancho: Number;
    CodigoBarras: string;
    CodigoInterno: string;
    DescripcionInventario: string;
    FechaRegistro: string;
    FechaRegistroOrdenCompra: string;
    IdKardex: Number;
    Largo: Number;
    Lote: string;
    OrdenCompra: string;
    PesoMaterial: Number;
    RazonSocialProveedor: string;
    Tipo: string;
    TotalM2: Number;
    NombreGrupo: string;
    NombreTipoInventario: string;
}